third_party/blink/web_tests/wpt_internal/document-transition/parsing/page-transition-tag-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Document Transitions: page-transition-tag with computed values</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://github.com/WICG/shared-element-transitions">
<meta name="assert" content="page-transition-tag supports custom keywords.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id=target></div>
<div id=scratch></div>
<script>

test_computed_value("page-transition-tag", "none");
test_computed_value("page-transition-tag", "foo");
test_computed_value("page-transition-tag", "bar");
test_computed_value("page-transition-tag", "baz");
</script>
</body>
</html>
